refactor: replace SimpleKeyItemBinding with SimpleItemUi

+++ app/src/main/java/com/osfans/trime/data/theme/model/GeneralStyle.kt
@@ -38,7 +38,7 @@ data class GeneralStyle(
     val keyFont: List<String>,
     val keyBorder: Int,
     val keyHeight: Int,
-    val keyLongTextSize: Int,
+    val keyLongTextSize: Float,
     val keyTextSize: Int,
     val keyTextOffsetX: Int,
     val keyTextOffsetY: Int,
